\name{RandomLettersGenerator}
\alias{RandomLettersGenerator}
\title{RandomLettersGenerator}
\usage{
library(RandomLettersGenerator)
randomLetterFunc(number)

The package "stringi" must be installed for this package to work
}
\description{
Generates a user-specified number of random letters
}
\examples{
library(RandomLettersGenerator)
randomLetterFunc(5)
}
